+++ application/src/test/java/com/ericsson/bss/cassandra/ecchronos/application/TestNoopStatementDecorator.java
@@ -18,7 +18,6 @@ import com.datastax.oss.driver.api.core.cql.Statement;
 import com.ericsson.bss.cassandra.ecchronos.application.config.Config;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.mockito.Mock;
 import org.mockito.junit.MockitoJUnitRunner;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -26,7 +25,6 @@ import static org.assertj.core.api.Assertions.assertThat;
 @RunWith(MockitoJUnitRunner.class)
 public class TestNoopStatementDecorator
 {
-    @Mock
     private Statement mockStatement;
 
     @Test
@@ -35,4 +33,5 @@ public class TestNoopStatementDecorator
         NoopStatementDecorator noopStatementDecorator = new NoopStatementDecorator(new Config());
         assertThat(noopStatementDecorator.apply(mockStatement)).isEqualTo(mockStatement);
     }
+
 }