+++ application/src/test/java/com/ericsson/bss/cassandra/ecchronos/application/config/TestRetryPolicy.java
@@ -20,38 +20,43 @@ import java.util.concurrent.TimeUnit;
 
 import org.junit.Before;
 import org.junit.Test;
-import org.mockito.Mockito;
+import org.junit.runner.RunWith;
+import org.mockito.junit.MockitoJUnitRunner;
 
+@RunWith(MockitoJUnitRunner.class)
 public class TestRetryPolicy {
 
-    RetryPolicy retryPolicy = Mockito.mock(RetryPolicy.class);
+    // No '@Mock' since RetryPolicy is under test
+    private RetryPolicy retryPolicy = new RetryPolicy();
 
     @Before
     public void setup()
     {
         retryPolicy.setUnit("seconds");
         retryPolicy.setDelay(5);
-        retryPolicy.setMaxAttempts(5);
         retryPolicy.setMaxDelay(30);
+        retryPolicy.setMaxAttempts(5);
     }
 
     @Test
-    public void testDefaultCurrentDelay() {
+    public void testDefaultCurrentDelay()
+    {
         long currentDelay = retryPolicy.currentDelay(1);
 
         assertEquals(TimeUnit.SECONDS.toMillis(10), currentDelay);
     }
 
     @Test
-    public void testMaxDelay() {
+    public void testMaxDelay()
+    {
         long currentDelay = retryPolicy.currentDelay(5);
 
         assertEquals(TimeUnit.SECONDS.toMillis(30), currentDelay);
     }
 
     @Test
-    public void testMaxDelayDisabled() {
-
+    public void testMaxDelayDisabled()
+    {
         retryPolicy.setMaxDelay(0);
 
         long currentDelay = retryPolicy.currentDelay(5);
@@ -60,7 +65,8 @@ public class TestRetryPolicy {
     }
 
     @Test
-    public void testDefaultConfigurationInMinutes() {
+    public void testDefaultConfigurationInMinutes()
+    {
         retryPolicy.setUnit("minutes");
         retryPolicy.setDelay(5);
         retryPolicy.setMaxAttempts(5);
@@ -80,4 +86,5 @@ public class TestRetryPolicy {
 
         assertEquals(TimeUnit.MINUTES.toMillis(160), currentDelay);
     }
+
 }