+++ core/src/main/java/com/ericsson/bss/cassandra/ecchronos/core/repair/types/RepairInfo.java
@@ -15,8 +15,8 @@
 
 package com.ericsson.bss.cassandra.ecchronos.core.repair.types;
 
-import javax.validation.constraints.Min;
-import javax.validation.constraints.NotBlank;
+import jakarta.validation.constraints.Min;
+import jakarta.validation.constraints.NotBlank;
 import java.util.List;
 import java.util.Objects;