+++ core/src/main/java/com/ericsson/bss/cassandra/ecchronos/core/repair/types/RepairStats.java
@@ -15,9 +15,9 @@
 
 package com.ericsson.bss.cassandra.ecchronos.core.repair.types;
 
-import javax.validation.constraints.Max;
-import javax.validation.constraints.Min;
-import javax.validation.constraints.NotBlank;
+import jakarta.validation.constraints.Max;
+import jakarta.validation.constraints.Min;
+import jakarta.validation.constraints.NotBlank;
 import java.util.Objects;
 
 @SuppressWarnings("VisibilityModifier")