+++ core/src/test/java/com/ericsson/bss/cassandra/ecchronos/core/scheduling/TestScheduleManager.java
@@ -153,20 +153,21 @@ public class TestScheduleManager
     public void testRunningTwoJobsInParallelShouldFail() throws InterruptedException
     {
         CountDownLatch job1Latch = new CountDownLatch(1);
-        TestJob job = new TestJob(ScheduledJob.Priority.HIGH, job1Latch);
-        CountDownLatch job2Latch = new CountDownLatch(1);
-        TestJob job2 = new TestJob(ScheduledJob.Priority.LOW, job2Latch);
-        myScheduler.schedule(job);
+        TestJob job1 = new TestJob(ScheduledJob.Priority.HIGH, job1Latch);
+        TestJob job2 = new TestJob(ScheduledJob.Priority.LOW, new CountDownLatch(1));
+
+        myScheduler.schedule(job1);
         myScheduler.schedule(job2);
 
         new Thread(() -> myScheduler.run()).start();
         new Thread(() -> myScheduler.run()).start();
-        waitForJobStarted(job);
-        job1Latch.countDown();
-        job2Latch.countDown();
-        waitForJobFinished(job);
 
-        assertThat(job.hasRun()).isTrue();
+        waitForJobStarted(job1);
+        assertThat(job2.hasStarted()).isFalse();
+        job1Latch.countDown();
+        waitForJobFinished(job1);
+
+        assertThat(job1.hasRun()).isTrue();
         assertThat(job2.hasRun()).isFalse();
         assertThat(myScheduler.getQueueSize()).isEqualTo(2);
     }